requirements.txt
python-dotenv
requests
google-cloud-storage
google-cloud-firestore
